backend/requirements.txt
uvicorn
fastapi
httpx
python-dotenv
boto3
aiosmtplib
slowapi
